Refactoring: сортировка по арх. FSD
