+++ src/features/getInitials/getInitials.ts
@@ -10,4 +10,3 @@ const getInitials: (name: string) => string = (name) => {
 };
 
 export default getInitials;
-//! удалить
