+++ src/features/index.ts
@@ -1,0 +1,3 @@
+import getInitials from "./getInitials/getInitials";
+
+export { getInitials };
