--- src/utils/index.ts
@@ -1,3 +1,0 @@
-import getInitials from "./getInitials";
-
-export { getInitials };
